Fix for types (int -> size_t)
git-svn-id: https://complement.svn.sourceforge.net/svnroot/complement/trunk@368 09cbc127-031f-0410-9985-c3ce8c6d183e

+++ explore/include/sockios/sockstream.cc
@@ -1,4 +1,4 @@
-// -*- C++ -*- Time-stamp: <99/09/02 21:50:39 ptr>
+// -*- C++ -*- Time-stamp: <99/09/14 20:51:04 ptr>
 
 #ident "$SunId$ %Q%"
 
@@ -441,7 +441,7 @@ int basic_sockbuf<charT, traits>::__rdsync()
 template<class charT, class traits>
 int basic_sockbuf<charT, traits>::recvfrom( void *buf, size_t n )
 {
-  int sz = sizeof( sockaddr_in );
+  size_t sz = sizeof( sockaddr_in );
 
   union {
       sockaddr_in inet;
